# openapi3.yaml
openapi: 3.0.1
info:
  title: Polygon Parts Geoserver API
  description: Unified REST API access to geoserver administration
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /workspaces:
    get:
      operationId: getAllWorkspacesNames
      tags:
        - workspaces
      summary: Get list of all workspace names
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/workspaces'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      operationId: createWorkspace
      tags:
        - workspaces
      summary: create new Workspace object by name
      requestBody:
        description: A json object containing desired workspace name
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: workspace name
      responses:
        201:
          description: Created
        409:
          description: Duplication Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /workspaces/{workspace}:
    get:
      operationId: getWorkspaceIfExist
      tags:
        - workspaces
      summary: Get Workspace object if exist by name or null
      parameters:
        - in: path
          name: workspace
          schema:
            type: string
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/workspaceInfo'
        404:
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      operationId: deleteWorkspace
      tags:
        - workspaces
      summary: delete Workspace by name
      parameters:
        - in: path
          name: workspace
          schema:
            type: string
          required: true
        - in: query
          name: isRecursive
          schema:
            type: boolean
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: 'OK'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              example:
                message: 'No such workspace'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              example:
                message: 'Workspace or related Namespace is not empty (and isRecursive not true)'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /workspaces/{workspace}/{newWorkspaceName}:
    put:
      operationId: updateWorkspace
      tags:
        - workspaces
      summary: Update Workspace object by name
      parameters:
        - in: path
          name: workspace
          schema:
            type: string
          required: true
        - in: path
          name: newWorkspaceName
          schema:
            type: string
          required: true
      responses:
        200:
          description: Modified
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              example:
                message: 'No such workspace'
        405:
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'

components:
  schemas:
    workspaces:
      type: array
      items:
        $ref: '#/components/schemas/workspace'
    workspace:
      type: object
      properties:
        name:
          type: string
        link:
          type: string
    workspaceInfo:
      type: object
      properties:
        name:
          type: string
        dateCreated:
          type: string
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string